requirements.txt
Flask
Flask-JWT-Extended
Flask-DotEnv
python-dotenv
